+++ android/anc/src/test/java/org/smartregister/fhircore/anc/data/patient/PatientRepositoryTest.kt
@@ -619,7 +619,7 @@ class PatientRepositoryTest : RobolectricTest() {
     coEvery { repository.revokeActiveStatusData(any()) } answers {}
     coEvery { fhirEngine.save(any()) } just runs
 
-    runBlocking { repository.deletePatient("99", DeletionReason.ENTRY_IN_ERROR) }
+    runBlocking { repository.deletePatient("99", DeletionReason.DIED) }
 
     val saveSlot = slot<Patient>()
 
@@ -640,7 +640,28 @@ class PatientRepositoryTest : RobolectricTest() {
     coEvery { repository.revokeActiveStatusData(any()) } answers {}
     coEvery { fhirEngine.save(any()) } just runs
 
-    runBlocking { repository.deletePatient("99", DeletionReason.MOVED_OUT) }
+    runBlocking { repository.deletePatient("99", DeletionReason.MOVED_AWAY) }
+
+    val saveSlot = slot<Patient>()
+
+    coVerify { fhirEngine.save(capture(saveSlot)) }
+
+    Assert.assertTrue(saveSlot.captured.active)
+    Assert.assertEquals(0, saveSlot.captured.link.size)
+  }
+
+  @Test
+  fun testDeletePatientWithOtherShouldRemovePatientLink() {
+    coEvery { fhirEngine.load(Patient::class.java, any()) } returns
+      Patient().apply {
+        active = true
+        addLink().other.reference = "ref"
+      }
+    coEvery { repository.revokeCarePlans(any(), any()) } answers {}
+    coEvery { repository.revokeActiveStatusData(any()) } answers {}
+    coEvery { fhirEngine.save(any()) } just runs
+
+    runBlocking { repository.deletePatient("99", DeletionReason.OTHER) }
 
     val saveSlot = slot<Patient>()
 
@@ -659,7 +680,7 @@ class PatientRepositoryTest : RobolectricTest() {
         meta.addTag(Coding("", "family", "family"))
       }
 
-    runBlocking { repository.deletePatient("99", DeletionReason.MOVED_OUT) }
+    runBlocking { repository.deletePatient("99", DeletionReason.MOVED_AWAY) }
   }
 
   @Test